+++ Makefile
@@ -39,6 +39,7 @@ BITCOIN_SRC :=					\
 	bitcoin/pullpush.c			\
 	bitcoin/script.c			\
 	bitcoin/shadouble.c			\
+	bitcoin/short_channel_id.c		\
 	bitcoin/signature.c			\
 	bitcoin/tx.c				\
 	bitcoin/varint.c
@@ -177,6 +178,7 @@ BITCOIN_HEADERS := bitcoin/address.h		\
 	bitcoin/pullpush.h			\
 	bitcoin/script.h			\
 	bitcoin/shadouble.h			\
+	bitcoin/short_channel_id.h		\
 	bitcoin/signature.h			\
 	bitcoin/tx.h				\
 	bitcoin/varint.h
